(fix) remove injective chain and exchange connectors
