+++ jest.config.js
@@ -15,7 +15,6 @@ module.exports = {
     'src/chains/avalanche/avalanche.ts',
     'src/chains/avalanche/pangolin/pangolin.ts',
     'src/chains/cosmos/cosmos.ts',
-    'src/chains/injective/injective.ts',
     'src/chains/near/near.ts',
     'src/chains/near/near.base.ts',
     'src/connectors/uniswap/uniswap.config.ts',
@@ -23,7 +22,6 @@ module.exports = {
     'src/connectors/uniswap/uniswap.lp.helper.ts',
     'src/connectors/openocean/openocean.ts',
     'src/connectors/pangolin/pangolin.ts',
-    'src/chains/injective/injective.mappers.ts',
     'src/connectors/quickswap/quickswap.ts',
     'src/connectors/sushiswap/sushiswap.ts',
     'src/connectors/traderjoe/traderjoe.ts',
