+++ package.json
@@ -32,13 +32,10 @@
     "@ethersproject/networks": "5.7.0",
     "@ethersproject/providers": "5.7.0",
     "@ethersproject/solidity": "5.7.0",
+    "@injectivelabs/sdk-ts": "^1.10.58",
     "@harmony-js/core": "^0.1.57",
     "@harmony-js/utils": "^0.1.56",
     "@improbable-eng/grpc-web": "^0.13.0",
-    "@injectivelabs/networks": "^1.10.8",
-    "@injectivelabs/sdk-ts": "^1.10.58",
-    "@injectivelabs/ts-types": "^1.0.29",
-    "@injectivelabs/wallet-ts": "^1.10.63",
     "@pancakeswap/sdk": "^2.4.5",
     "@pangolindex/sdk": "^1.1.0",
     "@perp/sdk-curie": "^1.16.0",
