+++ src/connectors/connectors.routes.ts
@@ -13,7 +13,6 @@ import { UniswapConfig } from './uniswap/uniswap.config';
 import { VVSConfig } from './vvs/vvs.config';
 import { RefConfig } from './ref/ref.config';
 import { PancakeSwapConfig } from './pancakeswap/pancakeswap.config';
-import { InjectiveCLOBConfig } from './injective/injective.clob.config';
 import { XsswapConfig } from './xsswap/xsswap.config';
 import { ConnectorsResponse } from './connectors.request';
 import { DexalotCLOBConfig } from './dexalot/dexalot.clob.config';
@@ -104,26 +103,6 @@ export namespace ConnectorsRoutes {
             trading_type: PancakeSwapConfig.config.tradingTypes,
             chain_type: PancakeSwapConfig.config.chainType,
             available_networks: PancakeSwapConfig.config.availableNetworks,
-          },
-          {
-            name: 'injective',
-            trading_type: InjectiveCLOBConfig.config.tradingTypes('spot'),
-            chain_type: InjectiveCLOBConfig.config.chainType,
-            available_networks: InjectiveCLOBConfig.config.availableNetworks,
-            additional_add_wallet_prompts: {
-              accountId:
-                'Enter your injective sub account id wallet key (input 0 if unsure) >>> ',
-            },
-          },
-          {
-            name: 'injective_perpetual',
-            trading_type: InjectiveCLOBConfig.config.tradingTypes('perp'),
-            chain_type: InjectiveCLOBConfig.config.chainType,
-            available_networks: InjectiveCLOBConfig.config.availableNetworks,
-            additional_add_wallet_prompts: {
-              accountId:
-                'Enter your injective sub account id wallet key (input 0 if unsure) >>> ',
-            },
           },
           {
             name: 'xswap',
