+++ src/services/connection-manager.ts
@@ -15,9 +15,6 @@ import { PancakeSwap } from '../connectors/pancakeswap/pancakeswap';
 import { Uniswap } from '../connectors/uniswap/uniswap';
 import { UniswapLP } from '../connectors/uniswap/uniswap.lp';
 import { VVSConnector } from '../connectors/vvs/vvs';
-import { InjectiveCLOB } from '../connectors/injective/injective';
-import { InjectiveClobPerp } from '../connectors/injective_perpetual/injective.perp';
-import { Injective } from '../chains/injective/injective';
 import {
   CLOBish,
   Ethereumish,
@@ -48,7 +45,6 @@ export type ChainUnion =
   | Cosmos
   | Ethereumish
   | Nearish
-  | Injective
   | Xdcish
   | Tezosish
   | Kujira;
@@ -63,8 +59,6 @@ export type Chain<T> = T extends Algorand
   ? Nearish
   : T extends Xdcish
   ? Xdcish
-  : T extends Injective
-  ? Injective
   : T extends Tezosish
   ? Tezosish
   : T extends KujiraCLOB
@@ -126,8 +120,6 @@ export async function getChainInstance(
     connection = BinanceSmartChain.getInstance(network);
   } else if (chain === 'xdc') {
     connection = Xdc.getInstance(network);
-  } else if (chain === 'injective') {
-    connection = Injective.getInstance(network);
   } else if (chain === 'tezos') {
     connection = Tezos.getInstance(network);
   } else if (chain === 'kujira') {
@@ -145,7 +137,6 @@ export type ConnectorUnion =
   | Perpish
   | RefAMMish
   | CLOBish
-  | InjectiveClobPerp
   | Tinyman
   | Plenty
   | Curve
@@ -161,8 +152,6 @@ export type Connector<T> = T extends Uniswapish
   ? RefAMMish
   : T extends CLOBish
   ? CLOBish
-  : T extends InjectiveClobPerp
-  ? InjectiveClobPerp
   : T extends Tinyman
   ? Tinyman
   : T extends Plenty
@@ -209,12 +198,8 @@ export async function getConnector<T>(
     connectorInstance = PancakeSwap.getInstance(chain, network);
   } else if (connector === 'sushiswap') {
     connectorInstance = Sushiswap.getInstance(chain, network);
-  } else if (chain === 'injective' && connector === 'injective_perpetual') {
-    connectorInstance = InjectiveClobPerp.getInstance(chain, network);
   } else if (chain === 'xdc' && connector === 'xsswap') {
     connectorInstance = Xsswap.getInstance(chain, network);
-  } else if (chain === 'injective' && connector === 'injective') {
-    connectorInstance = InjectiveCLOB.getInstance(chain, network);
   } else if (chain === 'avalanche' && connector === 'dexalot') {
     connectorInstance = DexalotCLOB.getInstance(network);
   } else if (chain == 'algorand' && connector == 'tinyman') {
