+++ src/templates/root.yml
@@ -92,10 +92,6 @@ configurations:
     configurationPath: pancakeswap.yml
     schemaPath: pangolin-schema.json
 
-  $namespace injective:
-    configurationPath: injective.yml
-    schemaPath: injective-schema.json
-
   $namespace xdc:
     configurationPath: xdc.yml
     schemaPath: ethereum-schema.json
