--- test-helpers/curl/requests/add_injective_key.json
@@ -1,6 +1,0 @@
-{
-  "privateKey": "$ETH_PRIVATE_KEY",
-  "chain": "injective",
-  "network": "mainnet",
-  "accountId": "0"
-}
