--- test-helpers/curl/requests/injective_balances.json
@@ -1,5 +1,0 @@
-{
-  "chain": "injective",
-  "network": "mainnet",
-  "address": "$INJECTIVE_SUBACCOUNTID"
-}
