--- test-helpers/curl/requests/injective_batch_create.json
@@ -1,22 +1,0 @@
-{
-  "chain": "injective",
-  "network": "mainnet",
-  "connector": "injective",
-  "address": "$INJECTIVE_SUBACCOUNTID",
-  "createOrderParams": [
-    {
-      "price": "2",
-      "amount": "0.10",
-      "side": "SELL",
-      "orderType": "LIMIT",
-      "market": "INJ-USDT"
-    },
-    {
-      "price": "3",
-      "amount": "0.10",
-      "side": "SELL",
-      "orderType": "LIMIT",
-      "market": "INJ-USDT"
-    }
-  ]
-}
