--- test-helpers/curl/requests/injective_perp_batch_create.json
@@ -1,24 +1,0 @@
-{
-  "chain": "injective",
-  "network": "mainnet",
-  "connector": "injective_perpetual",
-  "address": "$INJECTIVE_SUBACCOUNTID",
-  "createOrderParams": [
-    {
-      "price": "3.8",
-      "amount": "0.10",
-      "side": "BUY",
-      "orderType": "LIMIT_MAKER",
-      "market": "INJ-USDT",
-      "leverage": 1
-    },
-    {
-      "price": "3.85",
-      "amount": "0.10",
-      "side": "BUY",
-      "orderType": "LIMIT_MAKER",
-      "market": "INJ-USDT",
-      "leverage": 1
-    }
-  ]
-}
