--- test-helpers/curl/requests/injective_perp_batch_delete.json
@@ -1,16 +1,0 @@
-{
-  "chain": "injective",
-  "network": "mainnet",
-  "connector": "injective_perpetual",
-  "address": "$INJECTIVE_SUBACCOUNTID",
-  "cancelOrderParams": [
-    {
-      "market": "INJ-USDT",
-      "orderId": "0x3b433e730221666792ea249374a29f4b210497bbf9c0d6b785630265ead015a6"
-    },
-    {
-      "market": "INJ-USDT",
-      "orderId": "0x4f6d0a905a05210031e44f5e11f99190f0d3098e947eb7e76973900ffe3a593d"
-    }
-  ]
-}
