--- test-helpers/curl/requests/injective_perp_funding_info.json
@@ -1,6 +1,0 @@
-{
-  "chain": "injective",
-  "network": "mainnet",
-  "connector": "injective_perpetual",
-  "market": "INJ-USDT"
-}
