--- test-helpers/curl/requests/injective_perp_funding_payments.json
@@ -1,7 +1,0 @@
-{
-  "chain": "injective",
-  "network": "mainnet",
-  "connector": "injective_perpetual",
-  "market": "INJ-USDT",
-  "address": "$INJECTIVE_SUBACCOUNTID"
-}
