--- test-helpers/curl/requests/injective_perp_last_trade_price.json
@@ -1,6 +1,0 @@
-{
-  "chain": "injective",
-  "network": "mainnet",
-  "connector": "injective_perpetual",
-  "market": "INJ-USDT"
-}
