--- test-helpers/curl/requests/injective_perp_order_trades.json
@@ -1,8 +1,0 @@
-{
-  "chain": "injective",
-  "network": "mainnet",
-  "connector": "injective_perpetual",
-  "address": "$INJECTIVE_SUBACCOUNTID",
-  "market": "INJ-USDT",
-  "orderId": "0xf6f81a37796bd06a797484467302e4d6f72832409545e2e01feb86dd8b22e4b2"
-}
