--- test-helpers/curl/requests/injective_poll.json
@@ -1,5 +1,0 @@
-{
-  "chain": "injective",
-  "network": "mainnet",
-  "txHash": "6CDFBAAD3AAB9B1219CFA48A801229184DDC67B3750C9CE01690F2AB56E0FCE9"
-}
