--- test-helpers/curl/requests/injective_post_perp_order.json
@@ -1,12 +1,0 @@
-{
-  "chain": "injective",
-  "network": "mainnet",
-  "connector": "injective_perpetual",
-  "address": "$INJECTIVE_SUBACCOUNTID",
-  "price": "10",
-  "amount": "0.4",
-  "side": "BUY",
-  "market": "ATOM-USDT",
-  "leverage": 1,
-  "orderType": "LIMIT"
-}
