--- test-helpers/curl/requests/injective_transfer_to_bank.json
@@ -1,9 +1,0 @@
-{
-  "chain": "injective",
-  "network": "mainnet",
-  "connector": "injective",
-  "to": "$INJECTIVE_ACCOUNTID",
-  "from": "$INJECTIVE_SUBACCOUNTID",
-  "amount": "0.10",
-  "token": "USDT"
-}
